Various error handling/indentation/code structure fixes
