+++ arm9/source/virtual/vsaveexsv.h
@@ -8,6 +8,5 @@ u64 CheckVSaveDrive(void);
 
 bool ReadVSaveDir(VirtualFile* vfile, VirtualDir* vdir);
 int ReadVSaveFile(const VirtualFile* vfile, void* buffer, u64 offset, u64 count);
-bool VSaveIsExtData(void);
 
 u64 GetVSaveDriveSize(void);
